+++ notebooks/subdir/another_demo.clj
@@ -1,0 +1,3 @@
+(ns subdir.another-demo)
+
+(+ 1 2)