+++ CHANGELOG.md
@@ -15,15 +15,17 @@ All notable changes to this project will be documented in this file. This change
   Example `:quarto nil` will remove the default quarto configuration of theme and toc
 - show and continue for exceptions
 
-## Pending
+## [2-beta45 - 2025-05-22]
 - new option `:flatten-targets` (default true) do not create namespace folders.
-- new option `:sync-as-subdirs` (default true) includes the subdir itself (example "notebooks")
+- new option `:keep-sync-root` (default true) includes the subdir itself (example "notebooks")
 - syncing no longer creates empty directories
-- `make!` builds all files when given options `:base-source-path "notebooks"`, `:source-path nil`, and `:render true`
+- `make!` builds all files when given a non-nil `:base-source-path`, `:source-path nil`, and `:render true`
 - added `snippets/make-all!` to illustrate that pattern.
 - the CLI will now make all when given the `-r` render flag without any other arguments.
 - added a `:live-reload :toggle` option and changed the `snippets/watch!` to toggle.
 - The `:info` returned by `make!` is flatter and only reports things that happened.
+- Refactored the calls to Quarto, improved printing. Throw an error on Quaro failure (fixes #42).
+- Avoid syncing files when the source and target are the same.
 
 ## [2-beta44 - 2025-05-08]
 - updated deps (tools.cli, Charred, Hiccup, Kindly, Kindly-advice, tools.reader, babashka.fs, clojure-plus)